+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [![Downloads](https://img.shields.io/github/downloads/zevnda/steam-game-idler/1.7.12/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/zevnda/steam-game-idler/releases/download/1.7.12/Steam.Game.Idler_1.7.12_x64_en-US.msi)
 
 ## Changelog
-- #79: remove games with server-side achievements from achievement unlocker
+- #81: Re-added the `manual add` button for adding games that you do not own but have access to, such as family owned games
+  - Games can be manually added by doing to the Favorites List, clicking the `+` button and entering the game's ID